requirements.txt
# pNSFWMedia - NSFW Detection Model Requirements

# Core ML frameworks
tensorflow>=2.10.0
keras-tuner>=1.3.0
torch>=1.12.0
torchvision>=0.13.0

# CLIP (OpenAI)
# Install via: pip install git+https://github.com/openai/CLIP.git
ftfy
regex

# Data processing
numpy>=1.21.0
pandas>=1.3.0
Pillow>=9.0.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
tqdm>=4.64.0
tensorboard>=2.10.0

# Utilities
